+++ main.tf
@@ -82,6 +82,7 @@ resource "aviatrix_transit_gateway" "default" {
   ha_availability_domain           = local.ha_availability_domain
   ha_fault_domain                  = local.ha_fault_domain
   enable_multi_tier_transit        = var.enable_multi_tier_transit
+  enable_active_standby            = var.enable_active_standby
   enable_active_standby_preemptive = var.enable_active_standby_preemptive
   enable_s2c_rx_balancing          = var.enable_s2c_rx_balancing
   rx_queue_size                    = var.rx_queue_size