+++ versions.tf
@@ -2,8 +2,8 @@ terraform {
   required_providers {
     aviatrix = {
       source  = "aviatrixsystems/aviatrix"
-      version = "~>2.22.1"
+      version = "~>2.23.0"
     }
   }
-  required_version = ">= 1.1.0"
+  required_version = ">= 1.1.1"
 }